Refactor TweetCollection class to have inner classes for Document (#334)
* Refactor TweetCollection class to have inner classes for Document
* Refactor using Json Parser helper
* Refactor variable names
